Fix model timer_start.bpmn
The timer duration PT1D caused an invalid timer duration exception and
there were two start event tasks with the same ID present in the file.
Opened the file with the SpiffWorkflow editor and changed the duration
to P1D. The duplicated start event was automatically removed by the
editor.